-added cvs importer for Daum ergo bike -all graph data (distance, pulse,...) are now optional - this is not yet finished
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@381 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook.device.cm4xxm/src/net/tourbook/device/cm4xxm/CM4XXMDeviceReader.java
@@ -128,9 +128,7 @@ public class CM4XXMDeviceReader extends TourbookDevice {
 		return null;
 	}
 
-	public boolean processDeviceData(	String importFileName,
-										DeviceData deviceData,
-										HashMap<String, TourData> tourDataMap) {
+	public boolean processDeviceData(String importFileName, DeviceData deviceData, HashMap<String, TourData> tourDataMap) {
 
 		RandomAccessFile fileRawData = null;
 
@@ -358,7 +356,7 @@ public class CM4XXMDeviceReader extends TourbookDevice {
 
 							if (iData + 1 == iDataMax) {
 								// this is the last time slice
-								timeData.time = (short) (marker % CM4XXM_TIMESLICE);
+								timeData.time = marker % CM4XXM_TIMESLICE;
 							} else {
 								// this is a normal time slice
 								timeData.time = CM4XXM_TIMESLICE;
@@ -404,9 +402,7 @@ public class CM4XXMDeviceReader extends TourbookDevice {
 					tourData.setDeviceModeName(getDeviceModeName(profileId));
 
 					// set week of year
-					fCalendar.set(tourData.getStartYear(),
-							tourData.getStartMonth() - 1,
-							tourData.getStartDay());
+					fCalendar.set(tourData.getStartYear(), tourData.getStartMonth() - 1, tourData.getStartDay());
 					tourData.setStartWeek((short) fCalendar.get(Calendar.WEEK_OF_YEAR));
 				}
 
@@ -551,17 +547,17 @@ public class CM4XXMDeviceReader extends TourbookDevice {
 		if ((data & 0x0800) != 0) {
 			timeData.altitude = (short) (0xFFC0 | ((data & 0x0FC0) >> 6));
 			if (timeData.altitude < -16) {
-				timeData.altitude = (short) (-16 + ((timeData.altitude + 16) * 7));
+				timeData.altitude = (-16 + ((timeData.altitude + 16) * 7));
 			}
 		} else {
 			timeData.altitude = (short) ((data & 0x0FC0) >> 6);
 			if (timeData.altitude > 16) {
-				timeData.altitude = (short) (16 + ((timeData.altitude - 16) * 7));
+				timeData.altitude = (16 + ((timeData.altitude - 16) * 7));
 			}
 		}
 
 		// decode distance (6 bits)
-		timeData.distance = (data & 0x003F) * 10;
+		timeData.distance = (short) (data & 0x003F) * 10;
 
 	}
